requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# LLM and AI
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-groq==0.2.0
langchain-ollama==0.2.0
langsmith==0.1.99
openai==1.35.0
groq==0.9.0

# Email processing
imaplib2==3.6

# PDF processing
pdfplumber==0.11.0
PyMuPDF==1.24.0
pytesseract==0.3.10

# Data handling
pandas==2.2.0

# Document generation
fpdf==1.7.2

# Database
mysql-connector-python==8.3.0
